+++ src/data/actions/index.ts
@@ -1,5 +1,6 @@
 export * from './action';
 export * from './dashboard';
 export * from './root';
+export * from './settings';
 export * from './storage';
 export * from './ui';