+++ src/data/reducers/index.ts
@@ -1,5 +1,6 @@
 export * from './booted';
 export * from './dashboard';
+export * from './settings';
 export * from './storage';
 export * from './ui';
 export * from './version';