+++ src/plugins/core/widgets/time/Digital.tsx
@@ -1,25 +1,16 @@
-import { DateTime } from 'luxon';
 import * as React from 'react';
+import { FormattedTime } from 'react-intl';
 
 interface Props {
   hour12: boolean;
-  time: DateTime;
+  time: Date;
 }
-
-const TIME_12_SIMPLE = {
-  hour: 'numeric',
-  minute: '2-digit',
-  hour12: true,
-};
 
 const Digital: React.StatelessComponent<Props> = ({ time, hour12 }) => {
   return (
     <div className="Time Digital">
       <h1>
-        {time.toLocaleString(hour12
-          ? TIME_12_SIMPLE
-          : DateTime.TIME_24_SIMPLE
-        )}
+        <FormattedTime value={time} hour12={hour12} />
       </h1>
     </div>
   );