+++ src/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './getConvertedDate';
 export * from './readAsDataUrl';